// cargodelivery-front/generated/models/UserIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserIn
 */
export interface UserIn {
    /**
     * 
     * @type {string}
     * @memberof UserIn
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof UserIn
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof UserIn
     */
    middleName: string;
    /**
     * 
     * @type {string}
     * @memberof UserIn
     */
    roleId: string;
    /**
     * 
     * @type {string}
     * @memberof UserIn
     */
    phone: string;
    /**
     * 
     * @type {string}
     * @memberof UserIn
     */
    email: string;
}

/**
 * Check if a given object implements the UserIn interface.
 */
export function instanceOfUserIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "firstName" in value;
    isInstance = isInstance && "lastName" in value;
    isInstance = isInstance && "middleName" in value;
    isInstance = isInstance && "roleId" in value;
    isInstance = isInstance && "phone" in value;
    isInstance = isInstance && "email" in value;

    return isInstance;
}

export function UserInFromJSON(json: any): UserIn {
    return UserInFromJSONTyped(json, false);
}

export function UserInFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'middleName': json['middleName'],
        'roleId': json['roleId'],
        'phone': json['phone'],
        'email': json['email'],
    };
}

export function UserInToJSON(value?: UserIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'firstName': value.firstName,
        'lastName': value.lastName,
        'middleName': value.middleName,
        'roleId': value.roleId,
        'phone': value.phone,
        'email': value.email,
    };
}